fawkes_navgraph_node: provide navgraph-related services
This is based on recent additions to Fawkes' navgraph plugin. However,
we have a special constellation that makes a replication in this node
necessary: in the RCLL simulation, all robot-related data, including the
navgraph, is available on a per-robot rosocore. However, the central
integration with the planning system runs in a separate roscore.
Therefore, this additional node requires that the Fawkes instance writes
the generated graph files (upon generation) to a configurable location.
This node monitors that file for updates and then publishes the same
information and provides the same services on the central roscore.

This is similar to the reason for having the fawkes_skiller_node.