fawkes_navgraph_retriever: retrieve and save YAML navgraph
To avoid multi-master issues, we use fawkes_navgraph to read a navgraph
from a file and provide the same interface Fawkes' navgraph plugin does.
The new fawkes_navgraph_retriever can be used to access a remote
navgraph topic and write out a local file. This is necessary if the
Fawkes instance and the fawkes_navgraph node are separated onto multiple
machines (or VMs, containers, or Pods).